% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{addStyle}
\alias{addStyle}
\title{addStyle}
\usage{
addStyle(x, m, colorMap, pawn)
}
\arguments{
\item{x}{character vector containing HTML code}

\item{m}{character matrix containing table that is represented in x}

\item{colorMap}{character array of colors}

\item{pawn}{Boolean if TRUE use chess symbols rather than rectangles in html table}
}
\value{
returns modified HTML code
}
\description{
insert tags into HTML code to implement rotating table text
}
\examples{
v<-vector("character")
for(fun in funs_examples)
  v[fun]<-find_funs(fun)$package_name[1]

m<-concatPackFunc2(x_examples$funmat,v)
v2<-consolidate(v)
m2<-rearrangeM(m,v2)

m3<-removeZeroRowsCols(m2)

colors<-c("darkmagenta","darkolivegreen","darkorange3","brown4","red","blue")
colorMap<-mapFunctionsColors(m3[1,c(-1,-2)],m3[c(-1,-2),1],colors)

f<-sprintf("\%s/r2html",tempdir())
R2HTML::HTML(m3,file=f,append=FALSE)
x<-readLines(f)
y<-addStyle(x,m3,colorMap,pawn=TRUE)

}
