% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{attachedFunctions}
\alias{attachedFunctions}
\title{attachedFunctions}
\usage{
attachedFunctions(verbose)
}
\arguments{
\item{verbose}{Boolean if TRUE output several user messages}
}
\value{
returns a list whose components are
\itemize{
\item l list of user-selected packages and corresponding functions
\item where character vector of selected packages
\item prune character vector of selected functions
}
}
\description{
print a list of attached packages and
their functions for the user to select from
}
\examples{
if(interactive()){attachedFunctions(verbose=TRUE)}

}
