% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{attachedFunctionsBatch}
\alias{attachedFunctionsBatch}
\title{attachedFunctionsBatch}
\usage{
attachedFunctionsBatch(packs)
}
\arguments{
\item{packs}{list of character strings containing the names of packages
package name is like "pack", not like "package:pack"}
}
\value{
returns a list whose components are
\itemize{
\item l list of user-selected packages and corresponding functions
\item where character vector of selected packages
}
}
\description{
same as attachedFunctions() but no user interaction needed
}
\examples{
if(interactive()){
attachedFunctionsBatch(c("SherlockHolmes","textBoxPlacement"))
}

}
