% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{attachedPackages}
\alias{attachedPackages}
\title{attachedPackages}
\usage{
attachedPackages()
}
\value{
returns a character vector of selected packages
}
\description{
print a list of attached packages for the user to select from
}
\examples{
if(interactive()){attachedPackages()}

}
