% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{concatPackFunc2}
\alias{concatPackFunc2}
\title{concatPackFunc2}
\usage{
concatPackFunc2(m, v)
}
\arguments{
\item{m}{character matrix return value component $funmat of foodweb()}

\item{v}{character vector of package names returned by find_funz()}
}
\value{
returns augmented character matrix m
}
\description{
match the package names with the function names
}
\examples{
if(interactive()){
load("data/x_x.RData")
load("data/x_v.RData")
m<-concatPackFunc2(x_x$funmat,x_v)
}

}
