% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{find_funs}
\alias{find_funs}
\title{find_funs}
\usage{
find_funs(f)
}
\arguments{
\item{f}{name of function for which the package(s) are to be identified.}
}
\value{
returns dataframe with two columns:
package_name packages(s) which the function is part of (chr)
builtin_package  whether the package comes with standard R (a 'builtin'  package)
}
\description{
determine in which R package a function ‘resides’
copied and pasted from https://sebastiansauer.github.io/finds_funs/
}
\details{
copied and pasted from https://sebastiansauer.github.io/finds_funs/
}
\examples{
find_funs("cbind")

}
