% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{find_funz}
\alias{find_funz}
\title{find_funz}
\usage{
find_funz(packs, rfuns)
}
\arguments{
\item{packs}{list of character strings containing the names of the packages}

\item{rfuns}{list of character strings containing the names of functions in
packs to which the result is to be restricted}
}
\value{
returns vector of character strings, names are functions and values are packages
}
\description{
determine in which R package a function ‘resides’
}
\examples{
if(interactive()){
load("data/x_packages.RData")
load("data/x_funs.RData")
find_funz(packs=x_packages,rfuns=x_funs)
}

}
