% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{leftJustifyHack}
\alias{leftJustifyHack}
\title{leftJustifyHack}
\usage{
leftJustifyHack(v)
}
\arguments{
\item{v}{character vector}
}
\value{
returns a character vector whose components are padded with &nbs escape sequences
}
\description{
insert &nbs escape sequences into text string to pad html table data cell
}
\details{
we need to generate a set of character strings to insert into html table data cells.
we want the strings to be of equal length (matching the length of the longest string),
by left justifying the strings and padding them on the right with spaces.
}
\examples{
v<-vector("character")
for(fun in funs_examples)
  v[fun]<-find_funs(fun)$package_name[1]

v2<-leftJustifyHack(v)

}
