% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{mapFunctionsColors}
\alias{mapFunctionsColors}
\title{mapFunctionsColors}
\usage{
mapFunctionsColors(row1, col1, colors)
}
\arguments{
\item{row1}{character vector containing names of packages}

\item{col1}{character vector containing names of packages}

\item{colors}{character vector containing names of colors}
}
\value{
returns a character vector mapping colors to package names
}
\description{
map functions to color coding
}
\examples{
if(interactive()){
colors<-c("darkmagenta","darkolivegreen","darkorange3","brown4","red","blue")
load("data/x_m3.RData")
colorMap<-mapFunctionsColors(x_m3[1,c(-1,-2)],x_m3[c(-1,-2),1],colors)
}

}
