% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{rearrangeM}
\alias{rearrangeM}
\title{rearrangeM}
\usage{
rearrangeM(m, v2)
}
\arguments{
\item{m}{character matrix return value of concatPackFunc2()}

\item{v2}{list whose components are package names for permuting order of m,
return value of consolidate()}
}
\value{
returns rearranged version of m
}
\description{
rearrange the order of rows or columns of matrix based on entries in a vector
}
\examples{
if(interactive()){
load("data/x_m.RData")
load("data/x_v2.RData")
m2<-rearrangeM(x_m,x_v2)
}

}
