% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{removeZeroRowsCols}
\alias{removeZeroRowsCols}
\title{removeZeroRowsCols}
\usage{
removeZeroRowsCols(m)
}
\arguments{
\item{m}{character matrix whose entries are either "0" or "1"}
}
\value{
returns an altered version of character matrix m with removed rows and columns
}
\description{
delete rows and cols of matrix m that contain all "0"s
}
\examples{
if(interactive()){
load("data/x_m2.RData")
m3<-removeZeroRowsCols(x_m2)
}

}
