% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{rotStyle}
\alias{rotStyle}
\title{rotStyle}
\usage{
rotStyle()
}
\value{
returns character string containing html style definition for rotation
}
\description{
add html style definition for rotation
}
\examples{
r<-rotStyle()

}
