% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{rotTag}
\alias{rotTag}
\title{rotTag}
\usage{
rotTag(v1, v2, colorMap)
}
\arguments{
\item{v1}{character vector containing first row of matrix m (excluding first 2 entries of m)}

\item{v2}{character vector containing second row of matrix m (excluding first 2 entries of m)}

\item{colorMap}{character array of colors}
}
\value{
returns character vector containing inserted html tags
}
\description{
add html tag to rotate function name
}
\details{
see https://stackoverflow.com/questions/47261100/how-to-rotate-text-90-degrees-inline
also need to increase height of row to accommodate rotated text
see https://resultuniversity.com/html/html-table-width-height#:~:text=To\%20set\%20the\%20height\%20of\%20a\%20specific\%20row\%20in\%20an,property\%20in\%20pixels\%20or\%20percentages.
}
\examples{
v<-vector("character")
for(fun in funs_examples)
  v[fun]<-find_funs(fun)$package_name[1]

m<-concatPackFunc2(x_examples$funmat,v)
v2<-consolidate(v)
m2<-rearrangeM(m,v2)
m3<-removeZeroRowsCols(m2)
colors<-c("darkmagenta","darkolivegreen","darkorange3","brown4","red","blue")
colorMap<-mapFunctionsColors(m3[1,c(-1,-2)],m3[c(-1,-2),1],colors)
rt<-rotTag(m[1,c(-1,-2)],m[2,c(-1,-2)],colorMap)

}
