% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{addStyle}
\alias{addStyle}
\title{addStyle}
\usage{
addStyle(x, m, colorMap, pawn)
}
\arguments{
\item{x}{character vector containing HTML code}

\item{m}{character matrix containing table that is represented in x}

\item{colorMap}{character array of colors}

\item{pawn}{Boolean if TRUE use chess symbols rather than rectangles in html table}
}
\value{
returns modified HTML code
}
\description{
insert tags into HTML code to implement rotating table text
}
\examples{
if(interactive()){
load("data/x_x2.RData")
load("data/x_m5.RData")
load("data/x_colorMap.RData")
y<-addStyle(x_x2,x_m5,x_colorMap,pawn=TRUE)
}

}
