% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{colorTag}
\alias{colorTag}
\title{colorTag}
\usage{
colorTag(v1, v2, nc, x, colorMap, pawn)
}
\arguments{
\item{v1}{character vector first column of m (excluding first 2 entries of m)}

\item{v2}{character vector second column of m (excluding first 2 entries of m)}

\item{nc}{integer number of columns of m}

\item{x}{return value of replaceRotTag()}

\item{colorMap}{character array of colors}

\item{pawn}{Boolean if TRUE use chess symbols rather than rectangles in html table}
}
\value{
returns
}
\description{
add tag to color function name in column 2, based on package in column 1
}
\details{
v1 is first column of m (excluding first 2 entries of m)
v2 is second column of m (excluding first 2 entries of m)
}
