% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{consolidate}
\alias{consolidate}
\title{consolidate}
\usage{
consolidate(v)
}
\arguments{
\item{v}{character vector of package names component of return value of find_funz()}
}
\value{
returns a list whose components are character vector for permuting order of m
}
\description{
create a permutation list of package names for re-ordering
rows and columns of matrix m, in decreasing order of function counts per package
}
\examples{
if(interactive()){
load("data/x_v.RData")
l<-consolidate(x_v)
}

}
