% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{mapFunctionsColors}
\alias{mapFunctionsColors}
\title{mapFunctionsColors}
\usage{
mapFunctionsColors(row1, col1, colors)
}
\arguments{
\item{row1}{character vector containing names of packages}

\item{col1}{character vector containing names of packages}

\item{colors}{character vector containing names of colors}
}
\value{
returns a character vector mapping colors to package names
}
\description{
map functions to color coding
}
\examples{
v<-vector("character")
for(fun in funs_examples)
  v[fun]<-find_funs(fun)$package_name[1]

m<-concatPackFunc2(x_examples$funmat,v)
v2<-consolidate(v)

m2<-rearrangeM(m,v2)

m3<-removeZeroRowsCols(m2)

colors<-c("darkmagenta","darkolivegreen","darkorange3","brown4","red","blue")

colorMap<-mapFunctionsColors(m3[1,c(-1,-2)],m3[c(-1,-2),1],colors)

}
