% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{rearrangeM}
\alias{rearrangeM}
\title{rearrangeM}
\usage{
rearrangeM(m, v2)
}
\arguments{
\item{m}{character matrix return value of concatPackFunc2()}

\item{v2}{list whose components are package names for permuting order of m,
return value of consolidate()}
}
\value{
returns rearranged version of m
}
\description{
rearrange the order of rows or columns of matrix based on entries in a vector
}
\examples{
v<-vector("character")
for(fun in funs_examples)
  v[fun]<-find_funs(fun)$package_name[1]

m<-concatPackFunc2(x_examples$funmat,v)
v2<-consolidate(v)

m2<-rearrangeM(m,v2)

}
